// packages/yugilife-templates/src/generated-catalog.ts
export const generatedDefaultTemplateIds = { card: "card/series-10" } as const

export const generatedTemplateEntries = [
  {
    assets: {
      "card.series-10.artwork-mask": asset0_0,
      "card.series-10.pendulum-artwork-mask": asset0_1,
      "card.series-10.pendulum-frame-mask.spell": asset0_2,
      "card.series-10.pendulum-effect-mask.small": asset0_3,
      "card.series-10.pendulum-effect-mask.medium": asset0_4,
      "card.series-10.pendulum-effect-mask.large": asset0_5,
      "card.series-10.pendulum-effect-mask.large-lower": asset0_6,
      "card.series-10.pendulum-border.small": asset0_7,
      "card.series-10.pendulum-border.medium": asset0_8,
      "card.series-10.pendulum-border.large": asset0_9,
      "card.series-10.pendulum-scale.left": asset0_10,
      "card.series-10.pendulum-scale.right": asset0_11,
      "card.series-10.artwork-box": asset0_12,
      "card.series-10.attribute.dark": asset0_13,
      "card.series-10.attribute.divine": asset0_14,
      "card.series-10.attribute.earth": asset0_15,
      "card.series-10.attribute.fire": asset0_16,
      "card.series-10.attribute.light": asset0_17,
      "card.series-10.attribute.spell": asset0_18,
      "card.series-10.attribute.trap": asset0_19,
      "card.series-10.attribute.water": asset0_20,
      "card.series-10.attribute.wind": asset0_21,
      "card.series-10.border": asset0_22,
      "card.series-10.effects-border": asset0_23,
      "card.series-10.frame-texture": asset0_24,
      "card.series-10.link-arrows-mask": asset0_25,
      "card.series-10.link-arrow.active.bottom-center": asset0_26,
      "card.series-10.link-arrow.active.bottom-left": asset0_27,
      "card.series-10.link-arrow.active.bottom-right": asset0_28,
      "card.series-10.link-arrow.active.left-center": asset0_29,
      "card.series-10.link-arrow.active.right-center": asset0_30,
      "card.series-10.link-arrow.active.top-center": asset0_31,
      "card.series-10.link-arrow.active.top-left": asset0_32,
      "card.series-10.link-arrow.active.top-right": asset0_33,
      "card.series-10.link-arrow.inactive.bottom-center": asset0_34,
      "card.series-10.link-arrow.inactive.bottom-left": asset0_35,
      "card.series-10.link-arrow.inactive.bottom-right": asset0_36,
      "card.series-10.link-arrow.inactive.left-center": asset0_37,
      "card.series-10.link-arrow.inactive.right-center": asset0_38,
      "card.series-10.link-arrow.inactive.top-center": asset0_39,
      "card.series-10.link-arrow.inactive.top-left": asset0_40,
      "card.series-10.link-arrow.inactive.top-right": asset0_41,
      "card.series-10.link-texture": asset0_42,
      "card.series-10.xyz-texture": asset0_43,
      "card.series-10.spell-trap.icon.continuous": asset0_44,
      "card.series-10.spell-trap.icon.counter": asset0_45,
      "card.series-10.spell-trap.icon.equip": asset0_46,
      "card.series-10.spell-trap.icon.field": asset0_47,
      "card.series-10.spell-trap.icon.quick-play": asset0_48,
      "card.series-10.spell-trap.icon.ritual": asset0_49,
      "card.series-10.star.level": asset0_50,
      "card.series-10.star.rank": asset0_51,
      "font.itc-stone-serif-medium": asset0_52,
      "font.itc-stone-serif-medium-italic": asset0_53,
      "font.itc-stone-serif-std-bold": asset0_54,
      "font.stone-serif-itc-tt-sc-bold": asset0_55,
      "font.dfp-reisho-w6-jajwa": asset0_56,
      "font.matrix-bold-small-caps": asset0_57,
      "font.matrix-book": asset0_58,
      "font.matrix-regular-small-caps": asset0_59,
      "font.rogsansrfstd-bd": asset0_60,
      "font.stone-serif-itc-semi": asset0_61,
    },
    colorPresets: presets0,
    manifest: runtimeManifests[0],
    template: runtimeTemplates[0],
  },
] as const